pyproject.toml
[tool.poetry]
name = "station-backend"
version = "1.2.0"
description = ""
authors = ["Mikael <mikael.delacour@mikajaki.com>"]

[tool.poetry.dependencies]
python = "^3.6"
flask = "^1"
Flask-RESTful = "^0.3"
flask-cors = "^3"
pyjwt = "^1.7.1"
pywin32 = {'version'="225",  'platform'="win32"}
"infi.systray" = {'version'="0.1.12",  'platform'="win32"}

mjk-utils = {}
mjk-backend = {}
mjk-vnc-backend = {}
station-common = {}
eyelib-logger = {}

[tool.poetry.dev-dependencies]
pylint = "^2.4"
requests = "2.22"
openpyxl = "3"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
